+++ src/main/java/net/weg/taskmanager/repository/CardsPermissionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CardsPermissionRepository extends JpaRepository<CardsPermission, Integer> {
+public interface CardsPermissionRepository extends JpaRepository<CardsPermission, Long> {
 }
