+++ src/main/java/net/weg/taskmanager/repository/SelectRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SelectRepository extends JpaRepository<SelectProperty, Integer> {
+public interface SelectRepository extends JpaRepository<SelectProperty, Long> {
 }
