+++ src/main/java/net/weg/taskmanager/repository/SubTaskRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubTaskRepository extends JpaRepository<Subtask, Integer> {
+public interface SubTaskRepository extends JpaRepository<Subtask, Long> {
 }
