+++ src/main/java/net/weg/taskmanager/repository/UserTaskRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserTaskRepository extends JpaRepository<UserTask, UserTaskId> {
 
-    UserTask findByUserIdAndTaskId(Integer userId, Integer taskId);
+    UserTask findByUserIdAndTaskId(Long userId, Long taskId);
 
 }
